Google Photos Poster/requirements.txt
requests
selenium
beautifulsoup4
webdriver_manager
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
